Rename cmd package as command, add stubbed function in client.
